+++ src/backend/DeckManagement/DeckController.py
@@ -13,330 +13,98 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 """
 import gc
-import statistics
+import os
+import sys
 import threading
 import time
-# Import Python modules
-from concurrent.futures import ThreadPoolExecutor
-from copy import copy
-from dataclasses import dataclass
+from functools import lru_cache
 from queue import Queue
-from threading import Thread, Timer
+from threading import Thread
+from typing import TYPE_CHECKING
 
-import psutil
-from PIL import Image, ImageDraw, ImageFont, ImageSequence
+from PIL import Image
 from StreamDeck.Devices import StreamDeck
-from StreamDeck.Devices.StreamDeck import DialEventType, TouchscreenEventType
 from StreamDeck.Devices.StreamDeckPlus import StreamDeckPlus
+from StreamDeck.ImageHelpers import PILHelper
+from gi.repository import GLib
 from loguru import logger as log
 
-# Import own modules
-from src.backend.DeckManagement.BetterDeck import BetterDeck
-from src.backend.DeckManagement.HelperMethods import *
-from src.backend.DeckManagement.ImageHelpers import *
-from src.backend.DeckManagement.InputIdentifier import Input, InputEvent, InputIdentifier
-from src.backend.DeckManagement.Subclasses.ActionPermissionManager import ActionPermissionManager
-from src.backend.DeckManagement.Subclasses.FakeDeck import FakeDeck
-from src.backend.DeckManagement.Subclasses.KeyImage import InputImage
-from src.backend.DeckManagement.Subclasses.KeyLabel import KeyLabel
-from src.backend.DeckManagement.Subclasses.KeyLayout import ImageLayout
-from src.backend.DeckManagement.Subclasses.KeyVideo import InputVideo
-from src.backend.DeckManagement.Subclasses.ScreenSaver import ScreenSaver
-from src.backend.DeckManagement.Subclasses.SingleKeyAsset import SingleKeyAsset
-from src.backend.DeckManagement.Subclasses.background_video_cache import BackgroundVideoCache
-from src.backend.PageManagement.Page import ActionOutdated, Page, NoActionHolderFound
-
-process = psutil.Process()
-
-from gi.repository import GLib
-
-# Import signals
-from src.Signals import Signals
-
-# Import typing
-from typing import TYPE_CHECKING, ClassVar, cast
-
-from src.windows.mainWindow.elements.KeyGrid import KeyButton, KeyGrid
-from src.backend.PluginManager.ActionCore import ActionCore
-if TYPE_CHECKING:
-    from src.windows.mainWindow.elements.DeckStackChild import DeckStackChild
-    from src.backend.DeckManagement.DeckManager import DeckManager
-
-# Import globals
 import globals as gl
+from src.backend.DeckManagement.BetterDeck import BetterDeck
+from src.backend.DeckManagement.HelperMethods import recursive_hasattr
+from src.backend.DeckManagement.InputIdentifier import (
+    Input,
+    InputEvent,
+    InputIdentifier,
+)
+from src.backend.DeckManagement.Subclasses.Background import (
+    Background,
+    BackgroundImage,
+    BackgroundVideo,
+)
+from src.backend.DeckManagement.Subclasses.ControllerDial import (
+    ControllerDial,
+    ControllerDialState,
+)
+from src.backend.DeckManagement.Subclasses.ControllerInput import (
+    ControllerInput,
+    ControllerInputState,
+)
+from src.backend.DeckManagement.Subclasses.ControllerKey import (
+    ControllerKey,
+    ControllerKeyState,
+)
+from src.backend.DeckManagement.Subclasses.ControllerTouchScreen import (
+    ControllerTouchScreen,
+    ControllerTouchScreenState,
+)
+from src.backend.DeckManagement.Subclasses.FakeDeck import FakeDeck
+from src.backend.DeckManagement.Subclasses.InputStateManagers import (
+    BackgroundManager,
+    LabelManager,
+    LayoutManager,
+)
+from src.backend.DeckManagement.Subclasses.KeyGIF import KeyGIF
+from src.backend.DeckManagement.Subclasses.KeyLabel import KeyLabel
+from src.backend.DeckManagement.Subclasses.MediaPlayer import (
+    MediaPlayerSetImageTask,
+    MediaPlayerSetTouchscreenImageTask,
+    MediaPlayerTask,
+    MediaPlayerThread,
+)
+from src.backend.DeckManagement.Subclasses.ScreenSaver import ScreenSaver
+from src.backend.PageManagement.Page import Page
+from src.Signals import Signals
+from src.windows.mainWindow.elements.KeyGrid import KeyGrid
 
+if TYPE_CHECKING:
+    from src.backend.DeckManagement.DeckManager import DeckManager
+    from src.windows.mainWindow.elements.DeckStackChild import DeckStackChild
 
-@dataclass
-class MediaPlayerTask:
-    deck_controller: "DeckController"
-    page: Page
-    _callable: callable
-    args: tuple
-    kwargs: dict
+__all__ = [
+    "DeckController",
+    "Background",
+    "BackgroundImage",
+    "BackgroundVideo",
+    "BackgroundManager",
+    "ControllerDial",
+    "ControllerDialState",
+    "ControllerInput",
+    "ControllerInputState",
+    "ControllerKey",
+    "ControllerKeyState",
+    "ControllerTouchScreen",
+    "ControllerTouchScreenState",
+    "KeyGIF",
+    "KeyLabel",
+    "LabelManager",
+    "LayoutManager",
+    "MediaPlayerSetImageTask",
+    "MediaPlayerSetTouchscreenImageTask",
+    "MediaPlayerTask",
+    "MediaPlayerThread",
+]
 
-    def run(self):
-        self._callable(*self.args, **self.kwargs)
-
-@dataclass
-class MediaPlayerSetTouchscreenImageTask:
-    deck_controller: "DeckController"
-    page: Page
-    native_image: bytes
-
-    n_failed_in_row: ClassVar[dict] = {}
-
-    def run(self):
-        if not self.deck_controller.deck.is_touch():
-            return
-        try:
-            touchscreen_size = self.deck_controller.get_touchscreen_image_size()
-            self.deck_controller.deck.set_touchscreen_image(self.native_image, x_pos=0, y_pos=0, width=touchscreen_size[0], height=touchscreen_size[1]) # Maybe avoid to always merge the dial images before applying it
-            self.native_image = None
-            del self.native_image
-            MediaPlayerSetTouchscreenImageTask.n_failed_in_row = 0
-        except StreamDeck.TransportError as e:
-            log.error(f"Failed to set deck touchscreen image. Error: {e}")
-            MediaPlayerSetTouchscreenImageTask.n_failed_in_row += 1
-            if MediaPlayerSetTouchscreenImageTask.n_failed_in_row > 5:
-                log.debug(f"Failed to set touchscreen image for 5 times in a row for deck {self.deck_controller.serial_number()}. Removing controller")
-                
-                
-                self.deck_controller.deck.close()
-                self.deck_controller.media_player.running = False # Set stop flag - otherwise remove_controller will wait until this task is done, which it never will because it waits
-                gl.deck_manager.remove_controller(self.deck_controller)
-
-                gl.deck_manager.connect_new_decks()
-
-@dataclass
-class MediaPlayerSetImageTask:
-    deck_controller: "DeckController"
-    page: Page
-    key_index: int
-    native_image: bytes
-
-    n_failed_in_row: ClassVar[dict] = {}
-
-    def run(self):
-        try:
-            self.deck_controller.deck.set_key_image(self.key_index, self.native_image)
-            self.native_image = None
-            del self.native_image
-            MediaPlayerSetImageTask.n_failed_in_row[self.deck_controller.serial_number()] = 0
-        except StreamDeck.TransportError as e:
-            log.error(f"Failed to set deck key image. Error: {e}")
-
-            beta_resume = gl.settings_manager.get_app_settings().get("system", {}).get("beta-resume-mode", True)
-            if beta_resume:
-                return
-
-            MediaPlayerSetImageTask.n_failed_in_row[self.deck_controller.serial_number()] += 1
-            if MediaPlayerSetImageTask.n_failed_in_row[self.deck_controller.serial_number()] > 5:
-                log.debug(f"Failed to set key_image for 5 times in a row for deck {self.deck_controller.serial_number()}. Removing controller")
-                
-                
-                self.deck_controller.deck.close()
-                self.deck_controller.media_player.running = False # Set stop flag - otherwise remove_controller will wait until this task is done, which it never will because it waits
-                gl.deck_manager.remove_controller(self.deck_controller)
-
-                gl.deck_manager.connect_new_decks()
-
-
-class MediaPlayerThread(threading.Thread):
-    def __init__(self, deck_controller: "DeckController"):
-        super().__init__(name="MediaPlayerThread", daemon=True)
-        self.deck_controller: DeckController = deck_controller
-        self.FPS = 30 # Max refresh rate of the internal displays
-
-        self.running = False
-        self.media_ticks = 0
-
-        self.pause = False
-        self._stop = False
-
-        self.tasks: list[MediaPlayerTask] = []
-        self.image_tasks = {}
-        self.touchscreen_task = None
-        self._wake_event = threading.Event()
-
-        self.fps: list[float] = []
-        self.old_warning_state = False
-
-        self.show_fps_warnings = gl.settings_manager.get_app_settings().get("warnings", {}).get("enable-fps-warnings", True)
-
-    # @log.catch
-    def run(self):
-        self.running = True
-
-        while True:
-            start = time.time()
-
-            # self.check_connection()
-
-            if not self.pause:
-                has_bg_video = False
-
-                if self.deck_controller.background.video is not None:
-                    if self.deck_controller.background.video.page is self.deck_controller.active_page:
-                        has_bg_video = True
-                        # There is a background video
-                        video_each_nth_frame = self.FPS // self.deck_controller.background.video.fps
-                        if self.media_ticks % video_each_nth_frame == 0:
-                            self.deck_controller.background.update_tiles()
-
-                # Only iterate keys if there is animated content to update
-                if has_bg_video or self._needs_key_ticks():
-                    #TODO: generalize
-                    for key in self.deck_controller.inputs[Input.Key]:
-                        cast("ControllerKey", key).on_media_player_tick()
-
-                    for dial in self.deck_controller.inputs[Input.Dial]:
-                        cast("ControllerDial", dial).on_media_player_tick()
-                    # self.deck_controller.update_all_inputs()
-
-                # Perform media player tasks
-                self.perform_media_player_tasks()
-
-            self.media_ticks += 1
-
-            end = time.time()
-
-            # Use low FPS when idle (no animated content, no pending tasks)
-            has_pending = bool(self.tasks or self.image_tasks or self.touchscreen_task)
-            if has_pending or has_bg_video or getattr(self, '_cached_needs_ticks', False):
-                target_fps = self.FPS
-            else:
-                target_fps = 2  # Idle: just check for new tasks occasionally
-
-            self.append_fps(1 / (end - start))
-            self.update_low_fps_warning()
-            wait = max(0, 1/target_fps - (end - start))
-            if target_fps < self.FPS:
-                self._wake_event.wait(wait)
-                self._wake_event.clear()
-            else:
-                time.sleep(wait)
-
-            if self._stop:
-                break
-
-        self.running = False
-
-    def _needs_key_ticks(self) -> bool:
-        # Check once per second whether any key has animated content
-        if self.media_ticks % self.FPS != 0:
-            return getattr(self, '_cached_needs_ticks', False)
-        needs = False
-        for key in self.deck_controller.inputs.get(Input.Key, []):
-            state = key.get_active_state()
-            if state.key_video is not None:
-                needs = True
-                break
-        self._cached_needs_ticks = needs
-        return needs
-
-    def append_fps(self, fps: float) -> None:
-        self.fps.append(fps)
-        if len(self.fps) > self.FPS *2:
-            self.fps.pop(0)
-
-    def get_median_fps(self) -> float:
-        return statistics.median(self.fps)
-    
-    def update_low_fps_warning(self):
-        if not self.show_fps_warnings:
-            return
-        
-        show_warning = self.get_median_fps() < self.FPS * 0.8
-        if self.old_warning_state == show_warning:
-            return
-        self.old_warning_state = show_warning
-
-        self.set_banner_revealed(show_warning)
-
-
-    def set_show_fps_warnings(self, state: bool) -> None:
-        self.show_fps_warnings = state
-        if state:
-            self.old_warning_state = False
-        else:
-            self.set_banner_revealed(False)
-
-    def set_banner_revealed(self, state: bool) -> None:
-        deck_stack_child: "DeckStackChild" = self.deck_controller.get_own_deck_stack_child()
-        if deck_stack_child is None:
-            return
-        
-        # deck_stack_child.low_fps_banner.set_revealed(show_warning)
-        GLib.idle_add(deck_stack_child.low_fps_banner.set_revealed, state)
-
-
-    def stop(self) -> None:
-        self._stop = True
-        while self.running:
-            time.sleep(0.1)
-
-    def add_task(self, method: callable, *args, **kwargs):
-        self.tasks.append(MediaPlayerTask(
-            deck_controller=self.deck_controller,
-            page=self.deck_controller.active_page,
-            _callable=method,
-            args=args,
-            kwargs=kwargs
-        ))
-        self._wake_event.set()
-
-    def add_touchscreen_task(self, native_image: bytes):
-        self.touchscreen_task = MediaPlayerSetTouchscreenImageTask(
-            deck_controller=self.deck_controller,
-            page=self.deck_controller.active_page,
-            native_image=native_image
-        )
-        self._wake_event.set()
-
-    def add_image_task(self, key_index: int, native_image: bytes):
-        self.image_tasks[key_index] = MediaPlayerSetImageTask(
-            deck_controller=self.deck_controller,
-            page=self.deck_controller.active_page,
-            key_index=key_index,
-            native_image=native_image
-        )
-        self._wake_event.set()
-
-    def perform_media_player_tasks(self):
-        for task in self.tasks.copy():
-            if task.page is self.deck_controller.active_page:
-                task.run()
-
-            try:
-                self.tasks.remove(task)
-            except ValueError:
-                pass
-
-        for key in list(self.image_tasks.keys()):
-            try:
-                self.image_tasks[key].run()
-                del self.image_tasks[key]
-            except KeyError:
-                pass
-
-        if self.touchscreen_task is not None:
-            self.touchscreen_task.run()
-            del self.touchscreen_task
-            self.touchscreen_task = None
-    def check_connection(self):
-        try:
-            self.deck_controller.deck.get_firmware_version()
-        except StreamDeck.TransportError as e:
-            log.error(f"Seams like the deck is not connected. Error: {e}")
-            MediaPlayerSetImageTask.n_failed_in_row[self.deck_controller.serial_number()] += 1
-            if MediaPlayerSetImageTask.n_failed_in_row[self.deck_controller.serial_number()] > 5:
-                log.debug(f"Failed to contact the deck 5 times in a row: {self.deck_controller.serial_number()}. Removing controller")
-                
-                self.deck_controller.deck.close()
-                self.deck_controller.media_player.running = False # Set stop flat - otherwise remove_controller will wait until this task is done, which it never will because it waiuts
-                gl.deck_manager.remove_controller(self.deck_controller)
-
-                gl.deck_manager.connect_new_decks()
 
 class DeckController:
     def __init__(self, deck_manager: "DeckManager", deck: StreamDeck.StreamDeck):
@@ -662,6 +430,7 @@ class DeckController:
 
     @log.catch
     def load_all_inputs(self, page: Page, update: bool = True):
+        from concurrent.futures import ThreadPoolExecutor
         start = time.time()
         with ThreadPoolExecutor() as executor:
             futures = []
@@ -929,1973 +698,3 @@ class DeckController:
         except Exception as e:
             log.debug(f"Cougth dead deck error. Error: {e}")
             return False
-
-class Background:
-    def __init__(self, deck_controller: DeckController):
-        self.deck_controller = deck_controller
-
-        self.image = None
-        self.video = None
-
-        self.tiles: list[Image.Image] = [None] * deck_controller.deck.key_count()
-
-    def set_image(self, image: "BackgroundImage", update: bool = True) -> None:
-        self.image = image
-        if self.video is not None:
-            self.video.close()
-        self.video = None
-        gc.collect()
-
-        self.update_tiles()
-        if update:
-            self.deck_controller.update_all_inputs()
-
-    def set_video(self, video: "BackgroundVideo", update: bool = True) -> None:
-        if self.video is not None:
-            self.video.close()
-        self.image = None
-        self.video = video
-        gc.collect()
-
-        self.update_tiles()
-        if update:
-            self.deck_controller.update_all_inputs()
-
-    def set_from_path(self, path: str, fps: int = 30, loop: bool = True, update: bool = True, allow_keep: bool = True) -> None:
-        if path == "":
-            path = None
-        if path is None:
-            self.image = None
-            # self.video = None
-            self.set_video(None, update=False)
-            self.update_tiles()
-            if update:
-                self.deck_controller.update_all_inputs()
-        elif is_video(path):
-            if allow_keep:
-                if self.video is not None and self.video.video_path == path:
-                    self.video.page = self.deck_controller.active_page
-                    self.video.fps = fps
-                    self.video.loop = loop
-                    return
-            self.set_video(BackgroundVideo(self.deck_controller, path, loop=loop, fps=fps), update=update)
-        else:
-            if path is None:
-                return
-            if not os.path.isfile(path):
-                return
-            with Image.open(path) as image:
-                self.set_image(BackgroundImage(self.deck_controller, image.copy()), update=update)
-
-    def update_tiles(self) -> None:
-        old_tiles = self.tiles # Why store them and close them later? So that there is not key error if the media threads fetches them during the update
-        if self.image is not None:
-            self.tiles = self.image.get_tiles()
-        elif self.video is not None:
-            self.tiles = self.video.get_next_tiles()
-        else:
-            self.tiles = [self.deck_controller.generate_alpha_key() for _ in range(self.deck_controller.deck.key_count())]
-
-        for tile in old_tiles:
-            if tile is not None:
-                tile.close()
-                tile = None
-                del tile
-        del old_tiles
-
-class BackgroundImage:
-    def __init__(self, deck_controller: DeckController, image: Image) -> None:
-        self.deck_controller = deck_controller
-        self.image = image
-
-    def create_full_deck_sized_image(self) -> Image:
-        key_rows, key_cols = self.deck_controller.deck.key_layout()
-        key_width, key_height = self.deck_controller.get_key_image_size()
-        spacing_x, spacing_y = self.deck_controller.key_spacing
-
-        key_width *= key_cols
-        key_height *= key_rows
-
-        # Compute the total number of extra non-visible pixels that are obscured by
-        # the bezel of the StreamDeck.
-        spacing_x *= key_cols - 1
-        spacing_y *= key_rows - 1
-
-        # Compute final full deck image size, based on the number of buttons and
-        # obscured pixels.
-        full_deck_image_size = (key_width + spacing_x, key_height + spacing_y)
-
-        # Convert to RGBA first to preserve transparency, then resize
-        img_rgba = self.image.convert("RGBA")
-        return ImageOps.fit(img_rgba, full_deck_image_size, Image.LANCZOS)
-    
-    def crop_key_image_from_deck_sized_image(self, image: Image.Image, key):
-        deck = self.deck_controller.deck
-
-
-        key_rows, key_cols = deck.key_layout()
-        key_width, key_height = deck.key_image_format()['size']
-        spacing_x, spacing_y = self.deck_controller.key_spacing
-
-        # Determine which row and column the requested key is located on.
-        row = key // key_cols
-        col = key % key_cols
-
-        # Compute the starting X and Y offsets into the full size image that the
-        # requested key should display.
-        start_x = col * (key_width + spacing_x)
-        start_y = row * (key_height + spacing_y)
-
-        # Compute the region of the larger deck image that is occupied by the given
-        # key, and crop out that segment of the full image.
-        region = (start_x, start_y, start_x + key_width, start_y + key_height)
-        segment = image.crop(region)
-
-        # Return the segment directly, converting to RGBA to preserve transparency
-        return segment.convert("RGBA")
-    
-    def get_tiles(self) -> list[Image.Image]:
-        full_deck_sized_image = self.create_full_deck_sized_image()
-
-        tiles: list[Image.Image] = []
-        for key in range(self.deck_controller.deck.key_count()):
-            key_image = self.crop_key_image_from_deck_sized_image(full_deck_sized_image, key)
-            tiles.append(key_image)
-
-        return tiles
-
-class BackgroundVideo(BackgroundVideoCache):
-    def __init__(self, deck_controller: DeckController, video_path: str, loop: bool = True, fps: int = 30) -> None:
-        self.deck_controller = deck_controller
-        self.video_path = video_path
-        self.loop = loop
-        self.fps = fps
-
-        self.page: Page = self.deck_controller.active_page
-
-        self.active_frame: int = -1
-
-        super().__init__(video_path, deck_controller=deck_controller)
-
-    def get_next_tiles(self) -> list[Image.Image]:
-        # return [self.deck_controller.generate_alpha_key() for _ in range(self.deck_controller.deck.key_count())]
-        self.active_frame += 1
-
-        if self.active_frame >= self.n_frames:
-            if self.loop:
-                self.active_frame = 0
-
-        tiles =  self.get_tiles(self.active_frame)
-        try:
-            copied_tiles = [tile.copy() for tile in tiles]
-        except:
-            copied_tiles = [None for _ in range(len(tiles))]
-        return copied_tiles
-
-        frame = self.get_next_frame()
-        frame_full_sized_image = self.create_full_deck_sized_image(frame)
-
-        tiles: list[Image.Image] = []
-        for key in range(self.deck_controller.deck.key_count()):
-            key_image = self.crop_key_image_from_deck_sized_image(frame_full_sized_image, key)
-            tiles.append(key_image)
-
-        return tiles
-
-    def get_next_frame(self) -> Image.Image:
-        self.active_frame += 1
-
-        if self.active_frame >= self.n_frames:
-            if self.loop:
-                self.active_frame = 0
-        
-        return self.get_frame(self.active_frame)
-    
-    def create_full_deck_sized_image(self, frame: Image.Image) -> Image.Image:
-        key_rows, key_cols = self.deck_controller.deck.key_layout()
-        key_width, key_height = self.deck_controller.get_key_image_size()
-        spacing_x, spacing_y = self.deck_controller.key_spacing
-
-        key_width *= key_cols
-        key_height *= key_rows
-
-        # Compute the total number of extra non-visible pixels that are obscured by
-        # the bezel of the StreamDeck.
-        spacing_x *= key_cols - 1
-        spacing_y *= key_rows - 1
-
-        # Compute final full deck image size, based on the number of buttons and
-        # obscured pixels.
-        full_deck_image_size = (key_width + spacing_x, key_height + spacing_y)
-
-        # Resize the image to suit the StreamDeck's full image size. We use the
-        # helper function in Pillow's ImageOps module so that the image's aspect
-        # ratio is preserved.
-        return ImageOps.fit(frame, full_deck_image_size, Image.Resampling.HAMMING)
-    
-    def crop_key_image_from_deck_sized_image(self, image: Image.Image, key):
-        key_spacing = self.deck_controller.key_spacing
-        deck = self.deck_controller.deck
-
-
-        key_rows, key_cols = deck.key_layout()
-        key_width, key_height = deck.key_image_format()['size']
-        spacing_x, spacing_y = key_spacing
-
-        # Determine which row and column the requested key is located on.
-        row = key // key_cols
-        col = key % key_cols
-
-        # Compute the starting X and Y offsets into the full size image that the
-        # requested key should display.
-        start_x = col * (key_width + spacing_x)
-        start_y = row * (key_height + spacing_y)
-
-        # Compute the region of the larger deck image that is occupied by the given
-        # key, and crop out that segment of the full image.
-        region = (start_x, start_y, start_x + key_width, start_y + key_height)
-        segment = image.crop(region)
-
-        # Create a new key-sized image, and paste in the cropped section of the
-        # larger image.
-        key_image = PILHelper.create_key_image(deck)
-        key_image.paste(segment)
-
-        return key_image
-
-class KeyGIF(SingleKeyAsset):
-    def __init__(self, controller_key: "ControllerKey", gif_path: str, fps: int = 30, loop: bool = True):
-        super().__init__(controller_key)
-        self.gif_path = gif_path
-        self.fps = fps
-        self.loop = loop
-
-        self.active_frame: int = -1
-
-        self.gif = Image.open(self.gif_path)
-        self.frames = []
-        self.frame_delays = []
-        
-        # Extract frames and their delays
-        for frame in ImageSequence.Iterator(self.gif):
-            self.frames.append(frame.convert("RGBA"))
-            # Get frame delay from GIF metadata (in milliseconds)
-            # Default to 100ms (10fps) if no delay specified
-            delay = self.gif.info.get('duration', 100)
-            # Some GIFs use delay in centiseconds, convert to milliseconds
-            if delay < 50:
-                delay *= 10
-            self.frame_delays.append(delay)
-
-    def get_next_frame(self) -> Image.Image:
-        self.active_frame += 1
-
-        if self.active_frame >= len(self.frames):
-            if self.loop:
-                self.active_frame = 0
-            else:
-                self.active_frame = len(self.frames) - 1
-
-        return self.frames[self.active_frame]
-    
-    def get_frame_delay(self) -> float:
-        """Get delay for current frame in seconds"""
-        if self.active_frame < 0 or self.active_frame >= len(self.frame_delays):
-            return 1.0 / self.fps  # Fallback to fps-based timing
-        return self.frame_delays[self.active_frame] / 1000.0  # Convert ms to seconds
-    
-    def get_raw_image(self) -> Image.Image:
-        return self.get_next_frame()
-    
-    def close(self) -> None:
-        self.gif = None
-        self.frames = None
-        self.frame_delays = None
-        del self.gif
-        del self.frames
-        del self.frame_delays
-
-class LabelManager:
-    def __init__(self, controller_input: "ControllerInput"):
-        self.controller_input = controller_input
-        
-        self.page_labels = {}
-        self.action_labels = {}
-        self.scroll_wait = 25
-        self._has_scroll_labels_cache: bool = None
-
-        self.init_labels()
-        self.frames: dict[str, dict[str, int]] = {
-            "top": {
-                "position": 0,
-                "wait": self.scroll_wait
-            },
-            "center": {
-                "position": 0,
-                "wait": self.scroll_wait
-            },
-            "bottom": {
-                "position": 0,
-                "wait": self.scroll_wait
-            },
-        }
-
-    def init_labels(self):
-        for position in ["top", "center", "bottom"]:
-            self.page_labels[position] = KeyLabel(self.controller_input)
-            self.action_labels[position] = KeyLabel(self.controller_input)
- 
-    def clear_labels(self):
-        self.init_labels()
-        self._has_scroll_labels_cache = None
-
-    def set_page_label(self, position: str, label: "KeyLabel", update: bool = True):
-        if label is None:
-            label = self.page_labels[position]
-            label.clear_values()
-        else:
-            self.page_labels[position] = label
-
-        self._has_scroll_labels_cache = None
-        if update:
-            self.update_label(position)
-
-    @staticmethod
-    def _label_equals(a: "KeyLabel", b: "KeyLabel") -> bool:
-        return (a.text == b.text and a.font_size == b.font_size
-                and a.font_name == b.font_name and a.color == b.color
-                and a.font_weight == b.font_weight and a.style == b.style
-                and a.outline_width == b.outline_width
-                and a.outline_color == b.outline_color
-                and a.alignment == b.alignment)
-
-    def set_action_label(self, position: str, label: "KeyLabel", update: bool = True):
-        if label is None:
-            label = self.action_labels[position]
-            label.clear_values()
-        else:
-            old = self.action_labels.get(position)
-            if old is not None and self._label_equals(old, label):
-                return
-            self.action_labels[position] = label
-
-        self._has_scroll_labels_cache = None
-        GLib.idle_add(self.update_label_editor)
-        if update:
-            self.update_label(position)
-
-    def update_label_editor(self):
-        if not recursive_hasattr(gl, "app.main_win.sidebar.active_identifier"):
-            return
-        
-        if gl.app.main_win.sidebar.active_identifier != self.controller_input.identifier:
-            return
-        
-        controller = gl.app.main_win.get_active_controller()
-        if controller is not self.controller_input.deck_controller:
-            return
-
-        gl.app.main_win.sidebar.key_editor.label_editor.load_for_identifier(self.controller_input.identifier, self.controller_input.state)
-        
-
-    def get_use_page_label_properties(self, position: str) -> dict:
-        if self.page_labels.get(position) is None:
-            return {
-                "text": False,
-                "color": False,
-                "font-family": False,
-                "font-size": False,
-                "font-weight": False,
-                "font-style": False,
-                "outline_width": False,
-                "outline_color": False,
-                "alignment": False,
-            }
-        return {
-            "text": self.page_labels[position].text is not None,
-            "color": self.page_labels[position].color is not None,
-            "font-family": self.page_labels[position].font_name is not None,
-            "font-size": self.page_labels[position].font_size is not None,
-            "font-weight": self.page_labels[position].font_weight is not None,
-            "font-style": self.page_labels[position].style is not None,
-            "outline_width": self.page_labels[position].outline_width is not None,
-            "outline_color": self.page_labels[position].outline_color is not None,
-            "alignment": self.page_labels[position].alignment is not None,
-        }
-
-    def get_composed_label(self, position: str) -> str:
-        use_page_label_properties = self.get_use_page_label_properties(position)
-        
-        label = copy(self.action_labels.get(position)) or KeyLabel(self.controller_input)
-
-        # Set to page values
-        page_label = self.page_labels.get(position)
-        if page_label is not None:
-            if use_page_label_properties["text"]:
-                label.text = page_label.text
-            if use_page_label_properties["color"]:
-                label.color = page_label.color
-            if use_page_label_properties["font-family"]:
-                label.font_name = page_label.font_name
-            if use_page_label_properties["font-size"]:
-                label.font_size = page_label.font_size
-            if use_page_label_properties["font-weight"]:
-                label.font_weight = page_label.font_weight
-            if use_page_label_properties["font-style"]:
-                label.style = page_label.style
-            if use_page_label_properties["outline_width"]:
-                label.outline_width = page_label.outline_width
-            if use_page_label_properties["outline_color"]:
-                label.outline_color = page_label.outline_color
-            if use_page_label_properties["alignment"]:
-                label.alignment = page_label.alignment
-
-        injected = self.inject_defaults(label)
-        return self.fix_invalid(injected)
-    
-    def get_composed_labels(self) -> dict[str, "KeyLabel"]:
-        composed_labels = {}
-        for position in ["top", "center", "bottom"]:
-            composed_labels[position] = self.get_composed_label(position)
-        return composed_labels
-
-    
-    def inject_defaults(self, label: "KeyLabel"):
-        if label.text is None:
-            label.text = ""
-        if label.color is None:
-            label.color = gl.settings_manager.font_defaults.get("font-color") or (255, 255, 255, 255)
-        if label.font_name is None:
-            label.font_name = gl.settings_manager.font_defaults.get("font-family") or gl.fallback_font
-        if label.font_size is None:
-            label.font_size = round(gl.settings_manager.font_defaults.get("font-size") or 15)
-        if label.font_weight is None:
-            label.font_weight = round(gl.settings_manager.font_defaults.get("font-weight") or 400)
-        if label.style is None:
-            label.style = gl.settings_manager.font_defaults.get("font-style") or "normal"
-        if label.outline_width is None:
-            label.outline_width = round(gl.settings_manager.font_defaults.get("outline-width") or 2)
-        if label.outline_color is None:
-            label.outline_color = gl.settings_manager.font_defaults.get("outline-color") or (0, 0, 0, 255)
-        if label.alignment is None:
-            label.alignment = gl.settings_manager.font_defaults.get("alignment") or "center"
-
-        return label
-    
-    def fix_invalid(self, label: "KeyLabel"):
-        if not isinstance(label.text, str):
-            label.text = str(label.text)
-
-        return label
-
-    def update_label(self, position: str):
-        self.controller_input.update()
-
-    def get_available_width(self) -> int:
-        return self.controller_input.get_image_size()[0]
-
-    def get_has_scroll_labels(self) -> bool:
-        if self._has_scroll_labels_cache is not None:
-            return self._has_scroll_labels_cache
-
-        labels = self.get_composed_labels()
-        for label in labels:
-            if labels[label].text is not None and labels[label].text != "":
-                _, _, w, _ = labels[label].get_font().getbbox(labels[label].text)
-                if w > self.get_available_width():
-                    self._has_scroll_labels_cache = True
-                    return True
-        self._has_scroll_labels_cache = False
-        return False
-
-    def add_labels_to_image(self, image: Image.Image) -> Image.Image:
-        # image = image.rotate(self.deck.get_rotation()*-1)
-        draw = ImageDraw.Draw(image)
-
-        labels = self.get_composed_labels()
-        for label in labels:
-            text = labels[label].text
-            if text in [None, ""]:
-                continue
-
-            color = tuple(labels[label].color)
-            font = labels[label].get_font()
-            outline_width = labels[label].outline_width
-            outline_color = tuple(labels[label].outline_color)
-            alignment = labels[label].alignment
-
-            _, _, w, h = draw.textbbox((0, 0), text, font=font)
-
-            # Calculate x position based on alignment
-            padding = 3
-            if alignment == "left":
-                x_position = padding
-                anchor_x = "l"
-            elif alignment == "right":
-                x_position = image.width - padding
-                anchor_x = "r"
-            else:  # center (default)
-                x_position = image.width / 2
-                anchor_x = "m"
-
-            rolling_labels_enabled = gl.settings_manager.get_app_settings().get("general", {}).get("rolling-labels", True)
-            if rolling_labels_enabled and image.width < w:
-                # Need to scroll - always use center anchor for scrolling
-                start = image.width / 2 - (image.width - w) / 2 + 10
-                stop = image.width / 2 + (image.width - w) / 2 - 10
-
-                x_position = start - self.frames[label]["position"]
-                anchor_x = "m"
-                if x_position < stop:
-                    if self.frames[label]["wait"] == 0:
-                        x_position = start
-                        self.frames[label]["position"] = 0
-                        self.frames[label]["wait"] = self.scroll_wait
-                    else:
-                        self.frames[label]["wait"] -= 1
-                elif self.controller_input.media_ticks % 2 == 0:
-                    if self.frames[label]["wait"] == 0:
-                        if x_position == stop:
-                            self.frames[label]["wait"] = self.scroll_wait
-
-                        self.frames[label]["position"] += 1
-                    else:
-                        self.frames[label]["wait"] -= 1
-
-
-            if label == "top":
-                position = (x_position, h/2 + 3)
-            elif label == "bottom":
-                position = (x_position, image.height - h/2 - 3)
-            else:
-                position = (x_position, (image.height - 0) / 2)
-
-            # Use appropriate anchor based on alignment (x-anchor + "m" for vertical middle)
-            anchor = anchor_x + "m"
-
-            draw.text(position,
-                      text=text, font=font, anchor=anchor, align=alignment,
-                      fill=color, stroke_width=outline_width,
-                      stroke_fill=outline_color)
-
-        del draw
-
-        return image.copy()
-        # return image.copy().rotate(self.deck.get_rotation())
-
-
-class LayoutManager:
-    def __init__(self, controller_input: "ControllerInput"):
-        self.controller_input = controller_input
-
-        self.action_layout = ImageLayout()
-        self.page_layout = ImageLayout()
-
-    def clear(self):
-        self.action_layout = ImageLayout()
-        self.page_layout = ImageLayout()
-
-    def get_use_page_layout_properties(self) -> dict:
-        return {
-            "valign": self.page_layout.valign is not None,
-            "halign": self.page_layout.halign is not None,
-            "fill-mode": self.page_layout.fill_mode is not None,
-            "size": self.page_layout.size is not None
-        }
-    
-    def get_composed_layout(self) -> ImageLayout:
-        use_page_layout_properties = self.get_use_page_layout_properties()
-        
-        layout = copy(self.action_layout) or ImageLayout()
-
-        # Set to page values
-        page_layout = self.page_layout
-        if use_page_layout_properties["valign"]:
-            layout.valign = page_layout.valign
-        if use_page_layout_properties["halign"]:
-            layout.halign = page_layout.halign
-        if use_page_layout_properties["fill-mode"]:
-            layout.fill_mode = page_layout.fill_mode
-        if use_page_layout_properties["size"]:
-            layout.size = page_layout.size
-
-        return self.inject_defaults(layout)
-    
-    def inject_defaults(self, layout: ImageLayout):
-        if layout.valign is None:
-            layout.valign = 0
-        if layout.halign is None:
-            layout.halign = 0
-        if layout.fill_mode is None:
-            if isinstance(self.controller_input.identifier, Input.Key):
-                layout.fill_mode = "cover"
-            else:
-                layout.fill_mode = "contain"
-        if layout.size is None:
-            layout.size = 1
-
-        return layout
-    
-    def set_page_layout(self, layout: ImageLayout, update: bool = True):
-        self.page_layout = layout
-
-        if update:
-            self.update()
-
-    def set_action_layout(self, layout: ImageLayout, update: bool = True):
-        self.action_layout = layout
-
-        if update:
-            self.update()
-
-    def update(self):
-        self.controller_input.update()
-        GLib.idle_add(self.update_layout_editor)
-
-    def update_layout_editor(self):
-        if not recursive_hasattr(gl, "app.main_win.leftArea.deck_stack"):
-            return
-        
-        if gl.app.main_win.sidebar.active_identifier != self.controller_input.identifier:
-            return
-
-        controller = gl.app.main_win.get_active_controller()
-        if controller is not self.controller_input.deck_controller:
-            return
-
-        gl.app.main_win.sidebar.key_editor.image_editor.load_for_identifier(self.controller_input.identifier, self.controller_input.state)
-
-    def add_image_to_background(self, image: Image.Image, background: Image.Image) -> Image.Image:
-        if image is None:
-            return background
-        layout = self.get_composed_layout()
-
-        width, height = background.size
-        image_size = (int(width * layout.size), int(height * layout.size))
-
-        if 0 in image_size:
-            return background.copy()
-
-        if layout.fill_mode == "stretch":
-            image_resized = image.resize(image_size, Image.Resampling.HAMMING)
-        elif layout.fill_mode == "cover":
-            image_resized = ImageOps.cover(image, image_size, Image.Resampling.HAMMING)
-        else:
-            image_resized = ImageOps.contain(image, image_size, Image.Resampling.HAMMING)
-
-        halign = layout.halign
-        valign = layout.valign
-
-        left_margin = int((background.width - image_resized.width) * (halign + 1) / 2)
-        top_margin = int((background.height - image_resized.height) * (valign + 1) / 2)
-
-        # Create an image copy for the result
-        final_image = background.copy()
-
-        # Paste the resized foreground onto the composite image at the calculated position
-        if image_resized.has_transparency_data:
-            final_image.paste(image_resized, (left_margin, top_margin), image_resized)
-        else:
-            final_image.paste(image_resized, (left_margin, top_margin))
-
-        return final_image
-    
-
-class BackgroundManager:
-    def __init__(self, controller_input: "ControllerInput"):
-        self.controller_input = controller_input
-        
-        self.action_color: list[int] = None
-        self.page_color: list[int] = None
-
-    def set_action_color(self, color: list[int], update: bool = True) -> None:
-        self.action_color = color
-        if isinstance(color, list) and len(color) == 3:
-            self.action_color.append(255)
-
-        if update:
-            self.update()
-
-    def set_page_color(self, color: list[int], update: bool = True, update_ui: bool = True) -> None:
-        self.page_color = color
-        if isinstance(color, list) and len(color) == 3:
-            self.page_color.append(255)
-
-        if update:
-            self.update(ui=update_ui)
-
-    def update(self, ui: bool = True):
-        self.controller_input.update()
-        if ui:
-            GLib.idle_add(self.update_background_editor)
-
-    def update_background_editor(self):
-        if not recursive_hasattr(gl, "app.main_win.leftArea.deck_stack"):
-            return
-        
-        if gl.app.main_win.sidebar.active_identifier != self.controller_input.identifier:
-            return
-
-        controller = gl.app.main_win.get_active_controller()
-        if controller is not self.controller_input.deck_controller:
-            return
-
-        gl.app.main_win.sidebar.key_editor.background_editor.load_for_identifier(self.controller_input.identifier, self.controller_input.state)
-
-    def get_color_is_set(self, color: list[int]) -> bool:
-        return color not in [None, [None]*3, [None]*4]
-
-    def get_use_page_background(self) -> dict:
-        return self.get_color_is_set(self.page_color)
-    
-    def get_composed_color(self) -> list[int]:
-        if self.get_use_page_background() and self.get_color_is_set(self.page_color):
-            return self.page_color
-        elif self.get_color_is_set(self.action_color):
-            return self.action_color
-        else:
-            return [0] * 4
-
-
-class ControllerInputState:
-    def __init__(self, controller_input: "ControllerInput", state: int):
-        self.controller_input = controller_input
-        self.deck_controller = controller_input.deck_controller
-        self.state = state
-        self._overlay: Image.Image = None
-        self.hide_overlay_timer: Timer = None
-
-        # managers
-        self.layout_manager = LayoutManager(self.controller_input)
-        self.label_manager = LabelManager(self.controller_input)
-        self.background_manager = BackgroundManager(self.controller_input)
-
-        self.action_permission_manager = ActionPermissionManager(self)
-
-    def __int__(self):
-        return self.state
-    
-    def ready(self):
-        pass
-
-    def stop_overlay_timer(self):
-        if self.hide_overlay_timer is not None:
-            self.hide_overlay_timer.cancel()
-            self.hide_overlay_timer = None
-
-    def show_overlay(self, image: Image.Image, duration: int = -1):
-        """
-        duration: -1 for infinite
-        """
-        if duration == 0:
-            self.stop_overlay_timer()
-            self._overlay = None
-            self.update()
-        elif duration > 0:
-            self._overlay = image
-            self.update()
-            self.hide_overlay_timer = Timer(duration, self.hide_error)
-            self.hide_overlay_timer.start()
-        else:
-            self._overlay = image
-            self.update()
-
-    def hide_overlay(self):
-        self._overlay = False
-        self.update()
-
-    def show_error(self, duration: int = -1):
-        error_img = Image.open(os.path.join("Assets", "images", "error.png"))
-        self.show_overlay(error_img, duration=duration)
-
-    def hide_error(self):
-        self.hide_overlay()
-
-    def close_resources(self) -> None:
-        pass
-
-    def get_own_actions(self) -> list["ActionCore"]:
-        if not self.deck_controller.get_alive(): return []
-        active_page = self.deck_controller.active_page
-        active_page = self.controller_input.deck_controller.active_page
-        if active_page is None:
-            return []
-        if active_page.action_objects is None:
-            return []
-        actions = self.deck_controller.active_page.get_all_actions_for_input(self.controller_input.identifier, self.state)
-
-        return actions
-
-    def update(self) -> None:
-        if self.controller_input.state == self.state:
-            self.controller_input.update()
-    
-    def own_actions_update(self) -> None:
-        for action in self.get_own_actions():
-            if not isinstance(action, ActionCore):
-                continue
-            if not action.on_ready_called:
-                continue
-            action.on_update()
-
-    @log.catch
-    def own_actions_tick(self) -> None:
-        for action in self.get_own_actions():
-            if not isinstance(action, ActionCore):
-                continue
-            if not action.on_ready_called:
-                continue
-            action.on_tick()
-
-    @log.catch
-    def own_actions_event_callback(self, event: InputEvent, data: dict = None, show_notifications: bool = False) -> None:
-        for action in self.get_own_actions():
-            if isinstance(action, ActionOutdated):
-                if show_notifications:
-                    plugin_id = gl.plugin_manager.get_plugin_id_from_action_id(action.id)
-                    gl.app.send_outdated_plugin_notification(plugin_id)
-                continue
-            if isinstance(action, NoActionHolderFound):
-                if show_notifications:
-                    plugin_id = gl.plugin_manager.get_plugin_id_from_action_id(action.id)
-                    gl.app.send_missing_plugin_notification(plugin_id)
-                continue
-
-            # parsed_event = event
-            # if action.allow_event_configuration:
-                # parsed_event = action.event_manager.get_event_assigner_for_event(event)
-
-            if event is None:
-                continue
-
-            if not isinstance(action, ActionCore):
-                continue
-
-            action._raw_event_callback(event, data)
-
-    def own_actions_ready_threaded(self) -> None:
-        threading.Thread(target=self.own_actions_ready, name="own_actions_ready").start()
-
-    def own_actions_update_threaded(self) -> None:
-        threading.Thread(target=self.own_actions_update, name="own_actions_update").start()
-
-    def own_actions_tick_threaded(self) -> None:
-        threading.Thread(target=self.own_actions_tick, name="own_actions_tick").start()
-
-    def own_actions_event_callback_threaded(self, event: InputEvent, data: dict = None, show_notifications: bool = False) -> None:
-        threading.Thread(target=self.own_actions_event_callback, args=(event, data, show_notifications), name="own_actions_event_callback").start()
-
-    def remove_media(self) -> None:
-        page = self.controller_input.deck_controller.active_page
-        if page is None:
-            return
-
-        page.set_media_path(identifier=self.controller_input.identifier, state=self.state, path=None)
-
-        self.update()
-
-
-class ControllerInput:
-    def __init__(self, deck_controller: DeckController, state_class: ControllerInputState, identifier: InputIdentifier):
-        self.deck_controller = deck_controller
-        self.state = 0
-        self.hide_error_timer: Timer = None
-        self.hold_start_timer: Timer = None
-        self.ControllerStateClass = state_class
-        self.identifier: InputIdentifier = identifier
-        self.media_ticks: int = 0
-
-        self.is_visual: bool = True
-
-        self.enable_states: bool = True
-
-        self.states: dict[int, ControllerInputState] = {
-            0: self.ControllerStateClass(self, 0),
-        }
-
-        self.states[self.state].ready()
-
-    @staticmethod
-    def Available_Identifiers(deck):
-        raise AttributeError
-
-    def update(self) -> None:
-        pass
-
-    def event_callback(self) -> None:
-        pass
-
-    def start_hold_timer(self):
-        self.stop_hold_timer()
-
-        self.hold_start_timer = threading.Timer(self.deck_controller.hold_time, self.on_hold_timer_end)
-        self.hold_start_timer.setDaemon(True)
-        self.hold_start_timer.setName("HoldTimer")
-        self.hold_start_timer.start()
-
-    def stop_hold_timer(self):
-        if self.hold_start_timer is None:
-            return
-        
-        self.hold_start_timer.cancel()
-        self.hold_start_timer = None
-
-    def create_n_states(self, n: int):
-        if not self.enable_states:
-            n = 1
-
-        for state in self.states.values():
-            state.close_resources()
-        self.states.clear()
-
-        for i in range(n):
-            self.states[i] = self.ControllerStateClass(self, i)
-
-    def load_from_page(self, page: Page):
-        input_dict = self.identifier.get_config(page)
-        self.load_from_input_dict(input_dict)
-
-    def load_from_input_dict(self, page_dict, update: bool = True):
-        pass
-
-    def add_new_state(self, switch: bool = True):
-        if not self.enable_states:
-            if len(self.states) >= 1:
-                return
-            
-        d = self.identifier.get_config(self.deck_controller.active_page)
-
-        # Add new state
-        self.states[len(self.states)] = self.ControllerStateClass(self, len(self.states))
-        # Write to json
-        for state in self.states.keys():
-            d["states"].setdefault(str(state), {})
-
-        self.deck_controller.active_page.save()
-        gl.page_manager.update_dict_of_pages_with_path(self.deck_controller.active_page.json_path)
-
-        self.update_state_switcher()
-
-        if switch:
-            log.info(f"Switching to state: {len(self.states)-1}")
-            self.set_state(len(self.states)-1)
-
-    def remove_state(self, state: int):
-        d = self.identifier.get_config(self.deck_controller.active_page)
-
-        if str(state) in d["states"]:
-            d["states"].pop(str(state))
-
-        old_loaded_state = int(self.state)
-
-        state_to_remove = self.states.get(state)
-        if state_to_remove:
-            state_to_remove.close_resources()
-            self.states.pop(state)
-
-        # Fill gaps in self.states
-        sorted_state_keys = sorted(self.states.keys())
-
-        new_states = {}
-        state_map = {}
-        for new_key, old_key in enumerate(sorted_state_keys):
-            state_map[old_key] = new_key
-            self.states[old_key].state = new_key
-
-            if self.get_active_state() is self.states[old_key]:
-                self.state = new_key
-
-            new_states[new_key] = self.states[old_key]
-
-        self.states = new_states
-
-        new_states_dict = {}
-        for new_key, old_key in enumerate(d["states"].keys()):
-            new_states_dict[str(new_key)] = d["states"][old_key]
-
-        d["states"] = new_states_dict
-
-
-        self.deck_controller.active_page.save()
-        gl.page_manager.update_dict_of_pages_with_path(self.deck_controller.active_page.json_path)
-
-        self.update_state_switcher()
-
-        # Update - TODO: test
-        if state == self.state:
-            sort = sorted(list(self.states.keys()))
-            sort.reverse()
-            for s in sort:
-                if s <= state:
-                    self.set_state(s, allow_reload=True)
-                    break
-
-        gl.signal_manager.trigger_signal(Signals.RemoveState, state, state_map)
-
-    def update_state_switcher(self):
-        if gl.app.main_win.sidebar.active_identifier != self.identifier:
-            return
-
-        gl.app.main_win.sidebar.key_editor.state_switcher.set_n_states(len(self.states))
-
-    def get_active_state(self) -> "ControllerInputState":
-        return self.states.get(self.state, self.ControllerStateClass(self, -1))
-
-    def set_state(self, state: int, update_sidebar: bool = True, allow_reload: bool = False) -> None:
-        if state == self.state and not allow_reload:
-            return
-        
-        if state not in self.states:
-            log.error(f"Invalid state: {state}, must be one of {list(self.states.keys())}")
-            return
-        self.state = state
-
-        self.get_active_state().update()
-
-        if update_sidebar:
-            self.reload_sidebar()
-
-    def reload_sidebar(self) -> None:
-        visible_child = gl.app.main_win.leftArea.deck_stack.get_visible_child()
-        if visible_child is None:
-            return
-        controller = visible_child.deck_controller
-        if controller is None:
-            return
-        
-        if controller is not self.deck_controller:
-            return
-        if self.identifier != gl.app.main_win.sidebar.active_identifier:
-            return
-        
-        gl.app.main_win.sidebar.active_state = self.state
-        GLib.idle_add(gl.app.main_win.sidebar.update)
-
-    def load_from_config(self, config, update: bool = True):
-        n_states = len(config.get("states", {}))
-        self.create_n_states(max(1, n_states))
-
-        old_state_index = self.state
-
-        self.state = 0
-
-        #TODO: Reset states
-        for state in config.get("states", {}):
-            state: ControllerKeyState = self.states.get(int(state))
-            if state is None:
-                continue
-
-            state_dict = config["states"][str(state.state)]
-
-            self.get_active_state().own_actions_ready()
-            # state.own_actions_ready() # Why not threaded? Because this would mean that some image changing calls might get executed after the next lines which blocks custom assets
-
-            if update:
-                self.set_state(old_state_index)
-                self.update()
-
-    def clear(self, update: bool = True):
-        active_state = self.get_active_state()
-        active_state.clear()
-        if update:
-            self.update()
-
-    def has_unavailable_action(self) -> bool:
-        for action in self.get_active_state().get_own_actions():
-            if isinstance(action, ActionOutdated):
-                return True
-            if isinstance(action, NoActionHolderFound):
-                return True
-            
-        return False
-    
-    def get_empty_background(self) -> Image.Image:
-        pass
-
-    def get_image_size(self) -> tuple[int, int]:
-        pass
-
-class ControllerKey(ControllerInput):
-    def __init__(self, deck_controller: DeckController, ident: Input.Key):
-        super().__init__(deck_controller, ControllerKeyState, ident)
-        self.index = ident.get_index(deck_controller)
-        # Keep track of the current state of the key because self.deck_controller.deck.key_states seams to give inverted values in get_current_deck_image
-        self.press_state: bool = self.deck_controller.deck.key_states()[self.index]
-
-        self.down_start_time: float = None
-        
-        # GIF timing tracking
-        self.last_gif_update_time: float = 0
-
-    def on_hold_timer_end(self):
-        state = self.get_active_state()
-        state.own_actions_event_callback_threaded(
-            event=Input.Key.Events.HOLD_START
-        )
-
-    @staticmethod
-    def Available_Identifiers(deck):
-        return map(lambda x: f"{x[0]}x{x[1]}", map(lambda x: ControllerKey.Index_To_Coords(deck, x), range(deck.key_count())))
-
-    @staticmethod
-    def Index_To_Coords(deck, index):
-        rows, cols = deck.key_layout()    
-        y = index // cols
-        x = index % cols
-        return x, y
-    
-    @staticmethod
-    def Coords_To_Index(deck, coords):
-        if type(coords) == str:
-            coords = coords.split("x")
-        x, y = map(int, coords)
-        rows, cols = deck.key_layout()
-        return y * cols + x
-
-    def update(self, force: bool = False):
-        image = self.get_current_image()
-
-        # Quick hash check - skip expensive conversion if image unchanged
-        img_hash = hash(image.tobytes())
-        if not force and img_hash == getattr(self, '_last_img_hash', None):
-            image.close()
-            return
-        self._last_img_hash = img_hash
-
-        # Handle transparency properly - composite RGBA onto RGB to preserve smooth edges
-        if image.mode == "RGBA":
-            rgb_background = Image.new("RGB", image.size, (0, 0, 0))
-            rgb_background.paste(image, (0, 0), image)
-            rgb_image = rgb_background.rotate(self.deck_controller.deck.get_rotation())
-        else:
-            rgb_image = image.convert("RGB").rotate(self.deck_controller.deck.get_rotation())
-
-        if self.deck_controller.is_visual():
-            native_image = PILHelper.to_native_key_format(self.deck_controller.deck, rgb_image)
-            rgb_image.close()
-            self.deck_controller.media_player.add_image_task(self.index, native_image)
-
-        del rgb_image
-        self.set_ui_key_image(image)
-
-    def get_active_state(self) -> "ControllerKeyState":
-        return super().get_active_state()
-
-    def on_media_player_tick(self) -> None:
-        self.media_ticks += 1
-        current_time = time.time()
-
-        state = self.get_active_state()
-        needs_update = False
-        
-        # Check if we need to update based on content type
-        if state.key_video is not None:
-            if isinstance(state.key_video, KeyGIF):
-                # Use GIF frame delay timing
-                if self.last_gif_update_time == 0:
-                    self.last_gif_update_time = current_time
-                    needs_update = True
-                else:
-                    frame_delay = state.key_video.get_frame_delay()
-                    if current_time - self.last_gif_update_time >= frame_delay:
-                        self.last_gif_update_time = current_time
-                        needs_update = True
-            else:
-                # For non-GIF videos, use the original FPS-based logic
-                needs_update = True
-        elif self.deck_controller.background.video is not None or state.label_manager.get_has_scroll_labels():
-            # Other content types
-            needs_update = True
-
-        if needs_update:
-            self.update()
-
-    def event_callback(self, press_state):
-        screensaver_was_showing = self.deck_controller.screen_saver.showing
-        if press_state:
-            # Only on key down this allows plugins to control screen saver without directly deactivating it
-            self.deck_controller.screen_saver.on_key_change()
-        if screensaver_was_showing:
-            return
-        
-        self.deck_controller.mark_page_ready_to_clear(False)
-        self.press_state = press_state
-
-        self.update()
-
-        active_state = self.get_active_state()
-        if press_state: # Key down
-            self.down_start_time = time.time()
-            self.start_hold_timer()
-            active_state.own_actions_event_callback_threaded(
-                event=Input.Key.Events.DOWN,
-                show_notifications=True
-            )
-
-        elif self.down_start_time is not None: # Key up
-            if time.time() - self.down_start_time >= self.deck_controller.hold_time:
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Key.Events.HOLD_STOP
-                )
-            else:
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Key.Events.SHORT_UP
-                )
-            self.down_start_time = None
-            self.stop_hold_timer()
-            active_state.own_actions_event_callback_threaded(
-                event=Input.Key.Events.UP,
-                show_notifications=False
-            )
-        self.deck_controller.mark_page_ready_to_clear(True)
-
-    def get_current_image(self) -> Image.Image:
-        state = self.get_active_state()
-
-        background_color = self.get_active_state().background_manager.get_composed_color()
-
-        background: Image.Image = None
-        # Only load the background image if it's not gonna be hidden by the background color
-        if background_color[-1] < 255:
-            background = copy(self.deck_controller.background.tiles[self.index])
-
-        if background_color[-1] > 0:
-            background_color_img = Image.new("RGBA", self.deck_controller.get_key_image_size(), color=tuple(background_color))
-            
-            if background is None:
-                # Use the color as the only background - happens if background color alpha is 255
-                background = background_color_img
-            else:
-                background.paste(background_color_img, (0, 0), background_color_img)
-
-
-        if background is None:
-            background = self.deck_controller.generate_alpha_key().copy()
-
-        if state._overlay:
-            height = round(self.deck_controller.get_key_image_size()[1]*0.75)
-            img = state._overlay.resize((height, height))
-            background.paste(img, (int((self.deck_controller.get_key_image_size()[0] - height) // 2), int((self.deck_controller.get_key_image_size()[1] - height) // 2)), img)
-            return background
-
-
-        key_image: Image.Image = None
-        # rotation = self.deck_controller.get_deck_settings().get("rotation", {}).get("value", 0)
-        if state.key_image is not None:
-            image = state.key_image.get_raw_image()
-            key_image = state.layout_manager.add_image_to_background(
-                image=image,
-                background=background
-            )
-        elif state.key_video is not None:
-            image = state.key_video.get_raw_image()
-            key_image = state.layout_manager.add_image_to_background(
-                image=image,
-                background=background)
-        else:
-            key_image = background
-
-        labeled_image = state.label_manager.add_labels_to_image(key_image)
-
-        if self.is_pressed():
-            labeled_image = self.shrink_image(labeled_image)
-
-        if self.has_unavailable_action() and not self.deck_controller.screen_saver.showing:
-            labeled_image = self.add_warning_point(labeled_image)
-
-        if background is not None:
-            background.close()
-
-        key_image.close()
-
-        return labeled_image
-    
-    def add_warning_point(self, image: Image.Image, margin: int = 10, size: int = 10, color: tuple = (255, 150, 80)) -> Image.Image:
-        draw = ImageDraw.Draw(image)
-
-        # Calculate the coordinates of the top right circle
-        width, height = image.size
-        top_right_x = width - margin - size
-        top_right_y = margin
-
-        # Draw the circle
-        draw.ellipse((top_right_x, top_right_y, top_right_x + size, top_right_y + size), fill=color, outline=(0, 0, 0), width=2)
-
-        del draw
-        return image
-    
-
-    def is_pressed(self) -> bool:
-        return self.press_state
-    
-    def add_border(self, image: Image.Image) -> Image.Image:
-        image = image.copy()
-        draw = ImageDraw.Draw(image)
-        draw.rounded_rectangle((-1, -1, image.width, image.height), fill=None, outline=(255, 105, 0), width=8, radius=8)
-
-        return image
-
-    def shrink_image(self, image: Image.Image, factor: float = 0.7) -> Image.Image:
-        image = image.copy()
-        width = int(image.width * factor)
-        height = int(image.height * factor)
-        image = image.resize((width, height))
-
-        background = Image.new("RGBA", self.deck_controller.get_key_image_size(), (0, 0, 0, 0))
-
-        if image.has_transparency_data:
-            background.paste(image, (int((self.deck_controller.get_key_image_size()[0] - width) / 2), int((self.deck_controller.get_key_image_size()[1] - height) / 2)), image)
-        else:
-            background.paste(image, (int((self.deck_controller.get_key_image_size()[0] - width) / 2), int((self.deck_controller.get_key_image_size()[1] - height) / 2)))
-
-        image.close()
-
-        return background
-    
-    def load_from_input_dict(self, input_dict, update: bool = True, load_labels: bool = True, load_media: bool = True, load_background_color: bool = True):
-        """
-        Attention: Disabling load_media might result into disabling custom user assets
-        """
-        n_states = len(input_dict.get("states", {}))
-        self.create_n_states(max(1, n_states))
-
-        old_state_index = self.state
-
-        self.state = 0
-
-        #TODO: Reset states
-        for state in input_dict.get("states", {}):
-            state: ControllerKeyState = self.states.get(int(state))
-            if state is None:
-                continue
-
-            state_dict = input_dict["states"][str(state.state)]
-
-            ## Load media - why here? so that it doesn't overwrite the images chosen by the actions
-            if load_media:
-                state.key_image = None
-                state.key_video = None
-            
-            if load_labels:
-                state.label_manager.clear_labels()
-
-            # Reset action layout
-            layout = ImageLayout()
-            state.layout_manager.set_action_layout(layout, update=False)
-
-            state.own_actions_update() # Why not threaded? Because this would mean that some image changing calls might get executed after the next lines which blocks custom assets
-
-            ## Load labels
-            if load_labels:
-                for label in state_dict.get("labels", []):
-                    key_label = KeyLabel(
-                        controller_input=self,
-                        text=state_dict["labels"][label].get("text"),
-                        font_size=state_dict["labels"][label].get("font-size"),
-                        font_name=state_dict["labels"][label].get("font-family"),
-                        font_weight=state_dict["labels"][label].get("font-weight"),
-                        style=state_dict["labels"][label].get("style"),
-                        color=state_dict["labels"][label].get("color"),
-                        outline_width=state_dict["labels"][label].get("outline_width"),
-                        outline_color=state_dict["labels"][label].get("outline_color"),
-                        alignment=state_dict["labels"][label].get("alignment")
-                    )
-                    # self.add_label(key_label, position=label, update=False)
-                    state.label_manager.set_page_label(label, key_label, update=False)
-
-            ## Load media
-            if load_media:
-                path = state_dict.get("media", {}).get("path", None)
-                if path not in ["", None]:
-                    if is_image(path):
-                        with Image.open(path) as image:
-                            state.set_image(InputImage(
-                                controller_input=self,
-                                image=image.copy()
-                            ), update=False)
-                            
-                    elif is_svg(path):
-                        img = svg_to_pil(path, 192)
-                        state.set_image(InputImage(
-                            controller_input=self,
-                            image=img
-                        ), update=False)
-
-                    elif is_video(path):
-                        if os.path.splitext(path)[1].lower() == ".gif":
-                            state.set_video(KeyGIF(
-                                controller_key=self,
-                                gif_path=path,
-                                loop=state_dict.get("media", {}).get("loop", True),
-                                fps=state_dict.get("media", {}).get("fps", 30)
-                            )) # GIFs always update
-                        else:
-                            state.set_video(InputVideo(
-                                controller_input=self,
-                                video_path=path,
-                                loop = state_dict.get("media", {}).get("loop", True),
-                                fps = state_dict.get("media", {}).get("fps", 30),
-                            )) # Videos always update
-
-                layout = ImageLayout(
-                    fill_mode=state_dict.get("media", {}).get("fill-mode"),
-                    size=state_dict.get("media", {}).get("size"),
-                    valign=state_dict.get("media", {}).get("valign"),
-                    halign=state_dict.get("media", {}).get("halign"),
-                )
-                state.layout_manager.set_page_layout(layout, update=False)
-
-            elif len(state.get_own_actions()) > 1 and False: # Disabled for now - we might reuse it later
-                if state_dict.get("image-control-action") is None:
-                    with Image.open(os.path.join("Assets", "images", "multi_action.png")) as image:
-                        self.set_key_image(InputImage(
-                            controller_input=self,
-                            image=image.copy(),
-                        ), update=False)
-            
-            elif len(state.get_own_actions()) == 1:
-                if state_dict.get("image-control-action") is None:
-                    self.set_key_image(None, update=False)
-                # action = self.get_own_actions()[0]
-                # if action.has_image_control()
-
-            if load_background_color:
-                state.background_manager.set_page_color(state_dict.get("background", {}).get("color"), update=False)
-
-        if update:
-            self.set_state(old_state_index)
-            self.update()
-
-    def set_state(self, state: int, update_sidebar: bool = True, allow_reload: bool = False) -> None:
-        old_state = self.state
-        if state == old_state and not allow_reload:
-            return
-        super().set_state(state, False, allow_reload)
-        if update_sidebar:
-            self.reload_sidebar()
-
-    def set_ui_key_image(self, image: Image.Image) -> None:
-        if image is None:
-            return
-        
-        x, y = ControllerKey.Index_To_Coords(self.deck_controller.deck, self.index)
-
-        if self.deck_controller.get_own_key_grid() is None or not gl.app.main_win.get_mapped():
-            # Save to use later
-            self.deck_controller.ui_image_changes_while_hidden[self.identifier] = image # The ui key coords are in reverse order
-        else:
-            try:
-                GLib.idle_add(self.deck_controller.get_own_key_grid().buttons[x][y].set_image, image)
-            except:
-                print(f"Failed to set ui key image for {self.identifier}")
-        
-    def get_own_ui_key(self) -> KeyButton:
-        x, y = ControllerKey.Index_To_Coords(self.deck_controller.deck, self.index)
-        buttons = self.deck_controller.get_own_key_grid().buttons # The ui key coords are in reverse order
-        return buttons[x][y]
-    
-    def get_image_size(self) -> tuple[int, int]:
-        return self.deck_controller.get_key_image_size()
-
-class ControllerTouchScreen(ControllerInput):
-    def __init__(self, deck_controller: DeckController, ident: InputIdentifier):
-        super().__init__(deck_controller, ControllerTouchScreenState, ident)
-
-        self.enable_states = False
-
-    @staticmethod
-    def Available_Identifiers(deck):
-        if deck.is_touch():
-            return ["sd-plus"]
-        return []
-
-    def update(self) -> None:
-        image = self.get_current_image()
-        
-        # Touchscreen only supports JPEG, so composite RGBA onto background
-        if image.mode == "RGBA":
-            # Create a background image (black by default)
-            background = Image.new("RGB", image.size, (0, 0, 0))
-            # Composite the RGBA image onto the RGB background
-            background.paste(image, (0, 0), image)
-            image = background
-        
-        native_image = PILHelper.to_native_touchscreen_format(self.deck_controller.deck, image)
-        self.deck_controller.media_player.add_touchscreen_task(native_image)
-
-        self.set_ui_image(self.get_current_image())
-
-    def generate_empty_image(self) -> Image.Image:
-        return Image.new("RGBA", self.get_screen_dimensions(), (0, 0, 0, 0))
-    
-    def get_dial_image_area(self, identifier: Input.Dial) -> tuple[int, int, int, int]:
-        width, height = self.get_screen_dimensions()
-
-        n_dials = len(self.deck_controller.inputs[Input.Dial])
-        dial_index = identifier.index
-
-        start_x = int((dial_index / n_dials) * width)
-        start_y = 0
-        end_x = int(((dial_index + 1) / n_dials) * width)
-        end_y = height
-
-        return start_x, start_y, end_x, end_y
-    
-    def get_dial_image_area_size(self) -> tuple[int, int]:
-        width, height = self.get_screen_dimensions()
-
-        n_dials = len(self.deck_controller.inputs[Input.Dial])
-
-        return int(width / n_dials), height
-    
-    def get_empty_dial_image(self) -> Image.Image:
-        screen_width, screen_height = self.get_screen_dimensions()
-
-        n_dials = len(self.deck_controller.inputs[Input.Dial])
-
-        return Image.new("RGBA", (screen_width // n_dials, screen_height), (0, 0, 0, 0))
-
-    def set_ui_image(self, image: Image.Image) -> None:
-        if recursive_hasattr(self, "deck_controller.own_deck_stack_child.page_settings.deck_config.screenbar.image") and gl.app.main_win.get_mapped():
-            screenbar = self.deck_controller.own_deck_stack_child.page_settings.deck_config.screenbar
-            GLib.idle_add(screenbar.image.set_image, image)
-        else:
-            self.deck_controller.ui_image_changes_while_hidden[self.identifier] = image
-
-    def get_current_image(self) -> Image.Image:
-        active_state = self.get_active_state()
-        return active_state.get_current_image()
-
-    def event_callback(self, event_type, value):
-        screensaver_was_showing = self.deck_controller.screen_saver.showing
-        if event_type in (TouchscreenEventType.SHORT, TouchscreenEventType.LONG, TouchscreenEventType.DRAG):
-            self.deck_controller.screen_saver.on_key_change()
-        if screensaver_was_showing:
-            return
-        
-        active_state = self.get_active_state()
-        if event_type == TouchscreenEventType.DRAG:
-            # Check if from left to right or the other way
-            if value['x'] > value['x_out']:
-                active_state.own_actions_event_callback_threaded(
-                    Input.Touchscreen.Events.DRAG_LEFT
-                )
-            else:
-                active_state.own_actions_event_callback_threaded(
-                    Input.Touchscreen.Events.DRAG_RIGHT
-                )
-
-
-        #TODO get matching actions from the dials
-        elif event_type in (TouchscreenEventType.SHORT, TouchscreenEventType.LONG):
-            dial = self.get_dial_for_touch_x(value['x'])
-            if dial is not None:
-                dial_active_state = dial.get_active_state()
-                if dial_active_state is not None:
-
-                    event = Input.Dial.Events.SHORT_TOUCH_PRESS
-                    if event_type == TouchscreenEventType.LONG:
-                        event = Input.Dial.Events.LONG_TOUCH_PRESS
-
-                    dial_active_state.own_actions_event_callback_threaded(
-                        event,
-                        data={"x": value['x'], "y": value['y']},
-                        show_notifications=True
-                    )
-
-    def get_dial_for_touch_x(self, touch_x: float) -> "ControllerDial":
-        screen_width = self.deck_controller.get_touchscreen_image_size()[0]
-        n_dials = len(self.deck_controller.inputs[Input.Dial])
-        dial_index = int((touch_x / screen_width) * n_dials)
-
-        return self.deck_controller.get_input(Input.Dial(str(dial_index)))
-    
-    def get_screen_dimensions(self) -> tuple[int, int]:
-        return self.deck_controller.get_touchscreen_image_size()
-
-class ControllerDial(ControllerInput):
-    def __init__(self, deck_controller: DeckController, ident: InputIdentifier):
-        super().__init__(deck_controller, ControllerDialState, ident)
-
-        self.down_start_time: float = None
-
-    def on_hold_timer_end(self):
-        state = self.get_active_state()
-        state.own_actions_event_callback_threaded(
-            event=Input.Dial.Events.HOLD_START
-        )
-
-    def get_touch_screen(self) -> ControllerTouchScreen:
-        return self.deck_controller.get_input(Input.Touchscreen("sd-plus"))
-
-    @staticmethod
-    def Available_Identifiers(deck):
-        return map(str, range(deck.dial_count()))
-
-    def event_callback(self, event_type, value):
-        screensaver_was_showing = self.deck_controller.screen_saver.showing
-        if event_type == DialEventType.TURN:
-            self.deck_controller.screen_saver.on_key_change()
-        if event_type == DialEventType.PUSH and value:
-            # Only on push, not on hold to allow actions to enable the screensaver without directly causing it to wake up again
-            self.deck_controller.screen_saver.on_key_change()
-        if screensaver_was_showing:
-            return
-        
-        active_state = self.get_active_state()
-        if event_type == DialEventType.PUSH:
-            if value:
-                self.down_start_time = time.time()
-                self.start_hold_timer()
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Dial.Events.DOWN,
-                    show_notifications=True
-                )
-            elif self.down_start_time is not None:
-                self.stop_hold_timer()
-                if time.time() >= self.down_start_time + self.deck_controller.hold_time:
-                    active_state.own_actions_event_callback_threaded(
-                        event=Input.Dial.Events.HOLD_STOP
-                    )
-                else:
-                    active_state.own_actions_event_callback_threaded(
-                        event=Input.Dial.Events.SHORT_UP
-                    )
-                self.down_start_time = None
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Dial.Events.UP
-                )
-        
-        elif event_type == DialEventType.TURN:
-            if value < 0:
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Dial.Events.TURN_CCW
-                )
-            else:
-                active_state.own_actions_event_callback_threaded(
-                    event=Input.Dial.Events.TURN_CW
-                )
-
-    def load_from_input_dict(self, page_dict, update: bool = True):
-        n_states = len(page_dict.get("states", {}))
-        self.create_n_states(max(1, n_states))
-
-        old_state_index = self.state
-
-        self.state = 0
-
-        for state in page_dict.get("states", {}):
-            state: ControllerDialState = self.states.get(int(state))
-            if state is None:
-                continue
-
-            state_dict = page_dict["states"][str(state.state)]
-
-            # Reset action layout
-            layout = ImageLayout()
-            state.layout_manager.set_action_layout(layout, update=False)
-
-            state.own_actions_update() # Why not threaded? Because this would mean that some image changing calls might get executed after the next lines which blocks custom assets
-
-            ## Load labels
-            for label in state_dict.get("labels", []):
-                key_label = KeyLabel(
-                    controller_input=self,
-                    text=state_dict["labels"][label].get("text"),
-                    font_size=state_dict["labels"][label].get("font-size"),
-                    font_name=state_dict["labels"][label].get("font-family"),
-                    font_weight=state_dict["labels"][label].get("font-weight"),
-                    style=state_dict["labels"][label].get("style"),
-                    color=state_dict["labels"][label].get("color"),
-                    alignment=state_dict["labels"][label].get("alignment"),
-                )
-                state.label_manager.set_page_label(label, key_label, update=False)
-
-            ## Load media
-            path = state_dict.get("media", {}).get("path")
-            if path not in ["", None]:
-                if is_image(path):
-                    image = InputImage(
-                        controller_input=self,
-                        image=Image.open(path),
-                    )
-                    state.set_image(image, update=False)
-                elif is_svg(path):
-                    img = svg_to_pil(path, 192)
-                    state.set_image(InputImage(
-                        controller_input=self,
-                        image=img
-                    ), update=False)
-
-                elif is_video(path):
-                    if os.path.splitext(path)[1].lower() == ".gif":
-                        raise NotImplementedError("TODO") #TODO
-                        state.set_video(KeyGIF(
-                            controller_key=self,
-                            gif_path=path,
-                            loop=state_dict.get("media", {}).get("loop", True),
-                            fps=state_dict.get("media", {}).get("fps", 30)
-                        )) # GIFs always update
-                    else:
-                        state.set_video(InputVideo(
-                            controller_input=self,
-                            video_path=path,
-                            loop = state_dict.get("media", {}).get("loop", True),
-                            fps = state_dict.get("media", {}).get("fps", 30),
-                        )) # Videos always update
-
-            layout = ImageLayout(
-                fill_mode=state_dict.get("media", {}).get("fill-mode"),
-                size=state_dict.get("media", {}).get("size"),
-                valign=state_dict.get("media", {}).get("valign"),
-                halign=state_dict.get("media", {}).get("halign"),
-            )
-            state.layout_manager.set_page_layout(layout, update=False)
-
-            state.background_manager.set_page_color(state_dict.get("background", {}).get("color", [0, 0, 0, 0]), update=False)
-
-        if update:
-            self.set_state(old_state_index)
-            self.update()
-
-    def update(self):
-        if self.deck_controller.deck.is_touch():
-            self.get_touch_screen().update()
-
-    def get_active_state(self) -> "ControllerDialState":
-        return super().get_active_state()
-
-    def on_media_player_tick(self) -> None:
-        self.media_ticks += 1
-
-        state = self.get_active_state()
-        if not any([state.video, state.label_manager.get_has_scroll_labels()]):
-            return
-
-        self.update()
-
-    def get_image_size(self) -> tuple[int, int]:
-        if self.deck_controller.deck.is_touch():
-            return self.get_touch_screen().get_dial_image_area_size()
-        return (0, 0)
-    
-
-class ControllerTouchScreenState(ControllerInputState):
-    def __init__(self, controller_touch: "ControllerTouchScreen", state: int):
-        super().__init__(controller_touch, state)
-
-        self.controller_touch = controller_touch
-
-    def set_current_image(self, image: Image.Image):
-        self.current_image = image
-
-        self.update()
-
-    def get_current_image(self) -> Image.Image:
-        screen_width, screen_height = self.controller_touch.get_screen_dimensions()
-        
-        # Start with background image if set
-        background: Image.Image = None
-        active_page = self.controller_touch.deck_controller.active_page
-        background_image_path = active_page.get_background_image(
-            identifier=self.controller_touch.identifier, 
-            state=self.state
-        )
-        
-        if background_image_path and os.path.isfile(background_image_path):
-            try:
-                with Image.open(background_image_path) as img:
-                    # Resize to exact touchscreen dimensions (KISS - exact dimensions)
-                    background = ImageOps.fit(img, (screen_width, screen_height), Image.Resampling.LANCZOS).convert("RGBA")
-            except Exception as e:
-                log.error(f"Error loading background image: {e}")
-                background = None
-        
-        # Get background color from touchscreen state's background_manager
-        background_color = self.background_manager.get_composed_color()
-        
-        # If no background image, start with empty or colored background
-        if background is None:
-            # If background color has transparency (alpha < 255), start with transparent
-            if background_color[-1] < 255:
-                background = self.controller_touch.generate_empty_image()
-            
-            # If background color is set (alpha > 0), create colored background
-            if background_color[-1] > 0:
-                background_color_img = Image.new("RGBA", (screen_width, screen_height), color=tuple(background_color))
-                
-                if background is None:
-                    # Use the color as the only background - happens if background color alpha is 255
-                    background = background_color_img
-                else:
-                    # Paste color on top of transparent background
-                    background.paste(background_color_img, (0, 0), background_color_img)
-            
-            # If no background color was set, use empty image
-            if background is None:
-                background = self.controller_touch.generate_empty_image()
-        else:
-            # Background image exists - apply color overlay if set
-            if background_color[-1] > 0:
-                background_color_img = Image.new("RGBA", (screen_width, screen_height), color=tuple(background_color))
-                # Blend color over image
-                background = Image.alpha_composite(background, background_color_img)
-
-        # Paste dial images on top of the background
-        for dial in self.controller_touch.deck_controller.inputs[Input.Dial]:
-            state = dial.get_active_state()
-            image_area = self.controller_touch.get_dial_image_area(dial.identifier)
-            dial_image = state.get_rendered_touch_image()
-
-            background.paste(dial_image, image_area, dial_image)
-
-        return background
-
-
-    def update(self):
-        if self.controller_touch.get_active_state() is self:
-            self.controller_touch.update()
-
-    
-
-    def set_dial_image(self, identifier: Input.Dial, image: Image.Image, update: bool = True):
-        return
-        assert isinstance(identifier, Input.Dial)
-
-        area = self.get_dial_image_area(identifier)
-        width, height = area[2] - area[0], area[3] - area[1]
-
-        # Clear underground
-        empty_dial = self.get_empty_dial_image()
-        # Use alpha mask if empty_dial has transparency to prevent edge artifacts
-        if empty_dial.has_transparency_data:
-            self.current_image.paste(empty_dial, area, empty_dial)
-        else:
-            self.current_image.paste(empty_dial, area)
-
-        # Contain image into the area
-        image = ImageOps.contain(image, (width, height), Image.Resampling.HAMMING)
-
-        # Get x, y for centered position
-        x = area[0] + int((width - image.width) / 2)
-        y = area[1] + int((height - image.height) / 2)
-
-        self.current_image.paste(image, (x, y), image)
-
-        self.current_image.save("sd.png")
-
-        if update:
-            self.update()
-
-
-    def clear(self):
-        self.set_current_image(self.controller_touch.generate_empty_image())
-
-    def close_resources(self) -> None:
-        self.current_image.close()
-        del self.current_image
-
-class ControllerDialState(ControllerInputState):
-    def __init__(self, dial: "ControllerDial", state: int):
-        self.dial = dial
-
-        self.image: InputImage = None
-        self.video: InputVideo = None
-
-        self.touch_image: Image.Image = None
-
-        super().__init__(dial, state)
-
-    def set_image(self, image: "InputImage", update: bool = True) -> None:
-        if self.image is not None:
-            self.image.close()
-
-        self.image = image
-
-        if update:
-            self.update()
-
-    def set_video(self, video: "InputVideo") -> None:
-        if self.video is not None:
-            self.video.close()
-
-        self.video = video
-
-
-    def get_rendered_touch_image(self) -> Image.Image:
-        touch_screen = self.dial.get_touch_screen()
-
-        background: Image.Image = None
-
-        background_color = self.background_manager.get_composed_color()
-
-        if background_color[-1] < 255:
-            background = touch_screen.get_empty_dial_image()
-        if background_color[-1] > 0:
-            background_color_img = Image.new("RGBA", self.dial.get_image_size(), color=tuple(background_color))
-
-            if background is None:
-                # Use the color as the only background - happens if background color alpha is 255
-                background = background_color_img
-            else:
-                background.paste(background_color_img, (0, 0), background_color_img)
-        
-
-        image = None
-        if self.video is not None:
-            image = self.video.get_next_frame()
-        elif self.image is not None:
-            image = self.image.image
-
-        # rotation = self.deck_controller.get_deck_settings().get("rotation", {}).get("value", 0)
-
-        image = self.layout_manager.add_image_to_background(image, background)
-        image = self.label_manager.add_labels_to_image(image)
-
-        return image
-
-class ControllerKeyState(ControllerInputState):
-    def __init__(self, controller_key: "ControllerKey", state: int):
-        super().__init__(controller_key, state)
-
-        self.key_image: InputImage = None
-        self.key_video: InputVideo = None
-
-    def close_resources(self) -> None:
-        if self.key_image is not None:
-            self.key_image.close()
-            self.key_image = None
-        if self.key_video is not None:
-            self.key_video.close()
-            self.key_video = None
-            
-        # Reset GIF timing
-        if isinstance(self.controller_input, ControllerKey):
-            self.controller_input.last_gif_update_time = 0
-    
-    def set_image(self, key_image: "InputImage", update: bool = True) -> None:
-        if self.key_image is not None:
-            self.key_image.close()
-
-        self.key_image = key_image
-        self.key_video = None
-
-        if update:
-            self.update()
-
-    def set_video(self, key_video: "InputVideo") -> None:
-        self.key_video = key_video
-        if self.key_image is not None:
-            self.key_image.close()
-        self.key_image = None
-        
-        # Reset GIF timing for new video
-        if isinstance(self.controller_input, ControllerKey):
-            self.controller_input.last_gif_update_time = 0
-
-    def clear(self):
-        self.key_image = None
-        self.key_video = None
-        self.label_manager.clear_labels()
-        self.layout_manager.clear()
-        self.background_manager.set_page_color(None)
